requirements.txt
streamlit
numpy
pandas
scikit-learn
tensorflow
opencv-python
pillow
matplotlib
requests